Tests and Form components and backend architecture update
- Changed require to modern imports.
- Change server npm run dev to npm start

Moved the routes from index into

 /server/routes          - where we declare the routes
/server/controllers   - where we define the logic of the routes

CLIENT
Created an /actions folder - action creators, for api request to our backend
Created an /api folder - executes the axios requests with the function from /actions
Created an /reducer folder - which are pure functions that takes an action and the previous state of the application and returns the new state. (redux)

COMPONENTS
Form - used for creating a Test Document in the DB (still need to be able to append the questions in the question array.

Tests - Shows all the available tests
Test - An individual test.

SERVER
Created a test schema which references the question schemas in an array.